Unit Testing + Test Driven Updates + Bug Fixes
Need to finish WorkItemService and consider TestUtils + Utils before going too far ahead. Should also consider testing Workflow & Step services before jumping to engine, feeling i could change how it gets used (acknowledging that the direct methods could help)

1). Added/renamed WorkItemService to better support more transparent CRUD operations. Tests utilized JsonRepository<WorkItem>.

2).. Added a generic state summary, for ItemTasks and WorkItems. Need to consider validation.

3). Quite a few bugs like TaskState.mapToItemState.Done returning ERROR instead of Done,  some spelling mistakes in code, and remebering to cast Integer to int from summary map.

4). Dropped the dev Qualifers + Producers.

+++ tasktide/core/src/main/java/org/tasktide/core/model/workitem/WorkItem.java
@@ -423,11 +423,10 @@ public class WorkItem implements TaskTideModel {
     
     /**
      * Set task done count from summary
-     * 
      */
     public void setTaskDone() {
         Map<ItemState, Integer> data = this.summarizeByState();
-        this.taskDone = data.get(ItemState.DONE);
+        this.taskDone = (int) data.get(ItemState.DONE);
     }
